# Conflicts:
#	Makefile
#	include/mupdf/memento.h
#	include/mupdf/pdf/annot.h
#	platform/gl/gl-form.c
#	platform/win32/javaviewerlib.vcxproj
#	platform/win32/libextract.vcxproj
#	platform/win32/libmupdf.vcxproj
#	platform/win32/libthirdparty.vcxproj
#	platform/win32/mupdf.sln
#	platform/win32/mutool.vcxproj
#	source/fitz/filter-jbig2.c
#	source/fitz/load-jbig2.c
#	source/pdf/pdf-xref.c
#	thirdparty/extract

+++ source/fitz/xmltext-device.c
@@ -195,6 +195,7 @@ static void fz_xmltext_fill_image(fz_context *ctx, fz_device *dev_, fz_image *im
 		{
 			if (compressed->params.type == FZ_IMAGE_UNKNOWN)
 			{
+				/* unknown image type. */
 			}
 			else if (compressed->params.type == FZ_IMAGE_RAW)
 			{
@@ -336,7 +337,7 @@ static void fz_xmltext_fill_image(fz_context *ctx, fz_device *dev_, fz_image *im
 					fz_write_printf(ctx, dev->out, " ");
 					for (b=0; b<pixmap->n; ++b)
 					{
-						fz_write_printf(ctx, dev->out, "%02x", pixmap->samples[y*pixmap->stride + x*pixmap->n + b]);
+						fz_write_printf(ctx, dev->out, "%02x", pixmap->samples[y*(size_t)pixmap->stride + x*(size_t)pixmap->n + b]);
 					}
 				}
 				s_xml_endtag(ctx, dev->out, "line");
